QRCode/requirement.txt
numpy
pandas
Pillow
opencv-python
qrcode
pickle-mixin
notebook